Bump version to 1.4.4

+++ examples/output/xkb/scripts/install-xcompose.sh
@@ -7,6 +7,11 @@ layout=${1:-$layout}
 xcompose_file=${XCOMPOSEFILE:-$HOME/.XCompose}
 include_klfc="include \"%H/.XCompose-klfc-$layout\""
 
+if ! [ -f "$dir/../XCompose" ]; then
+  "$dir/uninstall-xcompose.sh" "$layout"
+  exit
+fi
+
 if ! [ -f "$xcompose_file" ]; then
   echo 'include "%L"' > "$xcompose_file"
 fi
@@ -15,4 +20,4 @@ if ! grep -qFx "$include_klfc" "$xcompose_file"; then
   echo "$include_klfc" >> "$xcompose_file"
 fi
 
-cp "$dir/XCompose" "$HOME/.XCompose-klfc-$layout"
+cp "$dir/../XCompose" "$HOME/.XCompose-klfc-$layout"
